Auto swap calls on hangup if holding
This patch adds a via to swap the calls
when we finish a call, and have calls holding,
to automatically swap to the call and release
the active one. Currently, when the active is
hanged up, the second keeps on hold, requiring
the user to press the pause button to resume
the holding one.

Signed-off-by: José Pekkarinen <jose.pekkarinen@foxhound.fi>