Make modem path and account path configurable
This patch makes the Telepathy account name and ofono modem path
configurable.

Configuring the modem path is required because not all devices have
modems named ril_*. The PinePhone has quectelqmi_* for example. The
ofono-on-ring mission-control plugin was created to address this issue,
but it is unable to handle oFono disconnecting and reconnecting to the
bus. I wasn't able to fix on-ring properly with a lot of time spent
looking at it, so this simpler patch was the answer instead.

Unfortunately, this plugin and the on-ring plugin do not use the same
account naming convention. Double unfortunately, our messaging
application is unable to handle a single message thread switching to a
differently-named account. It simply errors out if you try to send a
message in a thread created with the old account name. For that reason,
it is necessary to configure the account name just as much as the modem
path.

While I was in here, I noticed that we've got some buffer overflows
waiting to happen, so I solved those with GStrings where possible and
g_snprintf where not possible.